requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Vector stores and search
faiss-gpu>=1.7.4
rank-bm25>=0.2.2

# Document processing
langextract>=0.1.0
PyMuPDF>=1.23.0
pypdf>=3.17.0

# OpenSCENARIO 2.0 parser
py-osc2>=1.0.0

# Web framework
flask>=2.3.3
flask-cors>=4.0.0
flask-socketio>=5.3.6

# Data handling
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0

# Caching and storage
redis>=5.0.0
diskcache>=5.6.3

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
psutil>=5.9.0
ujson>=5.8.0

# Monitoring and logging
prometheus-client>=0.18.0
structlog>=23.2.0

# Optional: DeepSpeed for advanced GPU optimization
deepspeed>=0.12.0

# Development
pytest>=7.4.0
black>=23.9.0
